+++ packages/webcodecs/src/arraybuffer-to-uint8-array.ts
@@ -1,0 +1,5 @@
+export const arrayBufferToUint8Array = (
+	buffer: ArrayBuffer | null,
+): Uint8Array | null => {
+	return buffer ? new Uint8Array(buffer) : null;
+};
